requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0
pillow-heif==0.14.0
piexif==1.1.3

# File Monitoring
watchdog==3.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
